+++ arroyo-types/src/lib.rs
@@ -1120,3 +1120,4 @@ pub trait UdfContext: Sync {
     async fn init(&self) {}
     async fn close(&self) {}
 }
+
